requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
huggingface_hub
langchain-community
python-dotenv
langchain-cohere
unstructured
# --- New packages added below ---
requests
beautifulsoup4